extract MultipleTransactions class out of RangeSetTest
